% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Child Mortality vs. Total healthcare expenditure per capita}
\format{A data frame with 33,274 rows and 5 variables:
\describe{
  \item{country}{Country}
  \item{year}{Year}
  \item{health_exp}{Health Expenditure per capita (WDI)}
  \item{child_mort}{Child mortality (Gapminder)}
  \item{continent}{Continent}
}}
\source{
\url{https://ourworldindata.org/grapher/child-mortality-vs-health-expenditure?tab=data}
}
\usage{
mortality
}
\description{
Child Mortality (0-5 year-olds dying per 1,000 born) vs. Total healthcare
expenditure per capita (PPP).
}
\keyword{datasets}

