% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ourworldindata.R
\docType{package}
\name{ourworldindata}
\alias{ourworldindata}
\alias{ourworldindata-package}
\title{ourworldindata: An R data package containing tidy versions of datasets from
ourworldindata.org}
\description{
The ourworldindata package contains data frames which have been generated by
combining datasets from \href{https://ourworldindata.org}{OurWorldInData.org}: "an
online publication that shows how living conditions around the world are
changing". The data frames in this package have undergone tidying so that
they are suited to immediate analysis in R. The purpose of this package is to
serve as a central R resource for these datasets so that they might be used
for the likes of practice or exploratory data analysis in a replicable
manner.
}
\section{Data frames}{


The following data frames are loaded for use with the ourworldindata package:

\itemize{
  \item \code{\link{child_mortality}}
  \item \code{\link{financing_healthcare}}
}
}

